requirements.txt
numpy==1.26.0
pandas==2.2.2
requests==2.31.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
pymongo==4.8.0
plotly==5.24.0
dash==2.18.0
dash-bootstrap-components==1.6.0
setuptools>=68.0.0
scikit-learn==1.5.1
jupyter==1.1.1
dash==2.18.0
scikit-learn
joblib
psycopg[binary,pool]==3.2.1
sqlalchemy==2.0.33
pandas_ta